// src/features/media/ui/widgets/MediaWidget/types.ts
export interface SIRParams {
    scoreboard?: string;
    tabsPosition: string;
    matchId?: string | null;
    activeSwitcher?: string;
    forceTeamInvert?: boolean;
    layout?: string;
    pitchCustomBgColor?: string;
    customBrandColor?: string;
    disablePitchNoise?: boolean;
    disableWidgetHeader?: boolean;
    disablePeriods?: boolean;
}